feat: extract tool function
